% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLearningCurves.r
\name{plotLearningCurves}
\alias{plotLearningCurves}
\title{Utility function to plot learning curves of multiple models.}
\usage{
plotLearningCurves(models, labels, metric, ctrlFn, cv, colors, seed = 1)
}
\arguments{
\item{models}{{List}   - List of model objects.}

\item{labels}{{List}   - *Optional* List of labels associated with models}

\item{metric}{{String} - Metric models were trained with}

\item{cv}{{Data}   - Cross-validation data set (containing predictor)}

\item{colors}{{List}   - *Optional* List of colors associated with models}

\item{seed}{{Number} - Seed to use for training each model}
}
\description{
- Comparison helps determine if models contain bias or variance
  - Ideally want low training error and good generalization
  - Basic algorithm: split original data at different ratios
  - Re-train passed-in models on each split, and capture the error rate for
    both the train and valid data-sets.
  - NOTE: CV error is evaluated on _entire_ CV set, not subset
  - Plot the error (the score) versus the dataset size
}
\examples{
plotLearningCurves(
  models     = list(cfFit),
  metric     = "ROC",
  ctrlFn     = fitControl,
  cv         = valid,
  colors     = c("orange", "blue"),
  labels     = c("Glm Model 4", "Glm Model 6", "CForest"),
  seed       = SEED
)
}

