% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotROC.r
\name{plotROC}
\alias{plotROC}
\title{Utility function to plot multiple ROC curves against each other.}
\usage{
plotROC(models, res, predictor, cv, colors, titles, maxCurves = 3)
}
\arguments{
\item{models}{{List}   - List of model objects (Models expected to have
been trained with the "ROC" error metric.}

\item{res}{{String} - Response variable of each model}

\item{predictor}{{String} - Variable to predict in cv dataset}

\item{cv}{{Data}   - Cross-validation data set containing Predictor}

\item{colors}{{List}   - *Optional*; List of colors for each model}

\item{maxCurves}{{Number} - *Optional*; # curves per panel, defaults to 3.}

\item{labels}{{List}   - *Optional*; List of labels for each model}
}
\description{
- Comparison helps determine which models favour more TPs or less FPs
  - Curves are shown overlaid in the same panel upto a max threshold, upon
    when a new panel will be created.
}
\examples{
plotROC(
  models    = list(glmFit4, adaFit, rfFit, svmFit, cfFit),
  res       = "Foo",
  predictor = "Bar",
  cv        = valid,
  colors    = c("black", "orange", "yellow", "green", "red")
)
}

